refactor: consolidating the app service files
