--- azurerm/internal/services/web/app_service_certificate_validator.go
@@ -1,19 +1,0 @@
-package web
-
-import "fmt"
-
-// ValidateAppServiceCertificateID validates that the specified ID is a valid App Service Certificate ID
-func ValidateAppServiceCertificateID(i interface{}, k string) (warnings []string, errors []error) {
-	v, ok := i.(string)
-	if !ok {
-		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
-		return
-	}
-
-	if _, err := ParseAppServiceCertificateID(v); err != nil {
-		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
-		return
-	}
-
-	return warnings, errors
-}
