--- azurerm/internal/services/web/app_service_plan_validator.go
@@ -1,19 +1,0 @@
-package web
-
-import "fmt"
-
-// ValidateAppServicePlanID validates that the specified ID is a valid App Service Plan ID
-func ValidateAppServicePlanID(i interface{}, k string) (warnings []string, errors []error) {
-	v, ok := i.(string)
-	if !ok {
-		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
-		return
-	}
-
-	if _, err := ParseAppServicePlanID(v); err != nil {
-		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
-		return
-	}
-
-	return warnings, errors
-}
