+++ azurerm/internal/services/web/app_service_test.go
@@ -75,3 +75,49 @@ func TestParseAppService(t *testing.T) {
 		}
 	}
 }
+
+func TestValidateAppServiceID(t *testing.T) {
+	cases := []struct {
+		ID    string
+		Valid bool
+	}{
+		{
+			ID:    "",
+			Valid: false,
+		},
+		{
+			ID:    "nonsense",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Web",
+			Valid: false,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Web/sites/duckduckgo",
+			Valid: true,
+		},
+		{
+			ID:    "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/foo/providers/Microsoft.Web/Sites/duckduckgo",
+			Valid: false,
+		},
+	}
+
+	for _, tc := range cases {
+		t.Logf("[DEBUG] Testing Value %q", tc.ID)
+		_, errors := ValidateAppServiceID(tc.ID, "test")
+		valid := len(errors) == 0
+
+		if tc.Valid != valid {
+			t.Fatalf("Expected %t but got %t", tc.Valid, valid)
+		}
+	}
+}
