--- azurerm/internal/services/web/app_service_validator.go
@@ -1,19 +1,0 @@
-package web
-
-import "fmt"
-
-// ValidateAppServiceID validates that the specified ID is a valid App Service ID
-func ValidateAppServiceID(i interface{}, k string) (warnings []string, errors []error) {
-	v, ok := i.(string)
-	if !ok {
-		errors = append(errors, fmt.Errorf("expected type of %q to be string", k))
-		return
-	}
-
-	if _, err := ParseAppServiceID(v); err != nil {
-		errors = append(errors, fmt.Errorf("Can not parse %q as a resource id: %v", k, err))
-		return
-	}
-
-	return warnings, errors
-}
